CORDA-2572 Add `CounterFlowException` as cause of `FlowExceptions` from counterparty
To indicate that a `FlowException` has originated from a counterparty
the `cause` of the `FlowException` has been switched out to be a
`CounterFlowException`.

The `CounterFlowException` has information about the counterparty if
available.

The original cause of the `FlowException` is added to the
`CounterFlowException` so the extra information is not lost. On the
other hand, the stack trace is removed since it does not provide any
useful information about the exception to the user.